% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{annot}
\alias{annot}
\title{Annotation}
\usage{
annot(
  symbollist,
  species = NULL,
  ortholog = F,
  dboutput = "ncbi",
  idtype = NULL
)
}
\arguments{
\item{symbollist}{character list of IDs or Symbols}

\item{species}{character for species hs mm rn dr}

\item{ortholog}{logical return homo sapiens ortholog of species}

\item{dboutput}{character database used for Symbol annotation ncbi or ebi}

\item{idtype}{character annotation database ID type among SYMBOL (by defaut) GENE, ENST, ENSG, ENSP, UNIPROT}
}
\value{
data.frame
}
\description{
Annotate a list of symbols or IDs
}
\details{
supported is : symbol, ncbi gene, ensembl gene , transcrit, protein, uniprot swissrot, uniprot trembl
species : hs homo sapien , mm mus musculus , rn rattus norvegicus, dr danio rerio
}
\examples{
# not run
# annot(SymbolList)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
