% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot1.r
\name{boxplot1}
\alias{boxplot1}
\title{boxplot for one var}
\usage{
boxplot1(dat, ylab = "y", xlab = "TREATMENT", log = T)
}
\arguments{
\item{dat}{data.frame}

\item{ylab}{character}

\item{xlab}{character}

\item{log}{logical if TRUE data are delog in base 2}
}
\value{
plot
}
\description{
boxplot for one var
}
\examples{
# not run
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
