% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ena.r
\name{ena}
\alias{ena}
\title{MSigDB enrichment analysis}
\usage{
ena(
  SymbolList = NULL,
  geneannot = NULL,
  species = "hs",
  bg = 25000,
  filtergeneset = "all",
  overlapmin = 2,
  enaScoremin = 1,
  top = 80,
  labsize = 11,
  dpibarplot = "screen",
  path = ".",
  dirname = NULL
)
}
\arguments{
\item{SymbolList}{character Symbol or NCBI gene ID}

\item{geneannot}{data.frame}

\item{species}{character hs mm rn dr ss}

\item{bg}{numeric}

\item{filtergeneset}{regexp to filter geneset database}

\item{overlapmin}{numeric for minimum overlap between geneset and list}

\item{enaScoremin}{numeric for minimum ratio ena}

\item{top}{numeric top features to plot}

\item{labsize}{numeric size of function in barplot}

\item{dpibarplot}{character barplot resolution}

\item{path}{character for relative path of output directory}

\item{dirname}{character name for output}
}
\value{
file with enrichment analysis results
}
\description{
MSigDB enrichment analysis
}
\examples{
# not run
# ena( Symbollist , filtergeneset = "reactome")
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
