% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.r
\name{hc}
\alias{hc}
\title{Hierarchical clustering classification}
\usage{
hc(
  dat,
  factor = NULL,
  title = "Hierarchical Clustering",
  plot = TRUE,
  method = "complete",
  legendtitle = "TREATMENT",
  cexlabel = 0.6
)
}
\arguments{
\item{dat}{matrix numeric}

\item{factor}{factor}

\item{title}{character}

\item{plot}{logical}

\item{method}{character to choose agglomerative method of clustering dendrogramm}

\item{legendtitle}{character}

\item{cexlabel}{numeric}
}
\value{
a dendrogram
}
\description{
make a hierarchical clustering classification
}
\details{
Possible agglomerative method are the same hclust fonction : "complete" method by default
}
\examples{
# not run
# hc(dat)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
