% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.r
\name{heatmap}
\alias{heatmap}
\title{Heatmap}
\usage{
heatmap(
  dat,
  factor,
  method = "complete",
  dendrogram = "both",
  k = NULL,
  labCol = "",
  cexCol = 0.85,
  labRow = "",
  cexRow = 0.35,
  cexlegend = 0.65,
  keysize = 0.9,
  keycolor = c("darkgreen", "orange", "darkred"),
  parmar = c(5, 4, 5, 6)
)
}
\arguments{
\item{dat}{matrix numeric}

\item{factor}{factor}

\item{method}{character}

\item{dendrogram}{character to display 'none', 'row', 'column' or 'both' (by default) dendrograms}

\item{k}{numeric number of clusters to colorize for rows}

\item{labCol}{character}

\item{cexCol}{numeric}

\item{labRow}{Character}

\item{cexRow}{numeric}

\item{cexlegend}{numeric}

\item{keysize}{numeric}

\item{keycolor}{character of 3 for low mid high value of the key}

\item{parmar}{numeric 4 values for margin sizes}
}
\value{
no returned value
}
\description{
To make a heatmap
}
\details{
To make a heatmap from a matrix or a data.frame
}
\examples{
# not run
# library(magrittr)
# data(sif1)
# data(mat1)
# mat1 \%>\% heatmap(sif1$F3)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
