% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.r
\name{norm}
\alias{norm}
\title{Normalization}
\usage{
norm(dat, method = NULL, log = TRUE)
}
\arguments{
\item{dat}{data.frame}

\item{method}{character apply quantile normalization by default see details}

\item{log}{logical apply log base 2}
}
\value{
data.frame
}
\description{
quantile normalization and log2
}
\details{
for .method see limma normalizeBetweenArrays method
}
\examples{
# not run
# norm(dt)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
