% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic.r
\name{omic}
\alias{omic}
\title{Omic bioanalysis workflow}
\usage{
omic(
  dat,
  sif,
  annot = NULL,
  species = "hs",
  model = NULL,
  paired = NULL,
  nested = NULL,
  batch = NULL,
  addfactor = NULL,
  qcs = TRUE,
  threshold = c(c(2, 6, 11), c(2, 6, 11) + 120),
  padj = "none",
  pattern = TRUE,
  venn = TRUE,
  cluster = TRUE,
  nc = c(2, 3, 6, 12),
  heatmap = TRUE,
  maxheatmap = NULL,
  volcanoplot = TRUE,
  lineplot = TRUE,
  boxplotrow = TRUE,
  ena = TRUE,
  enamin = 2,
  filtergeneset = "all",
  enaScoremin = 1.1,
  bg = 25000,
  sample = NULL,
  dopar = NULL,
  path = ".",
  dirname = NULL,
  zip = FALSE,
  remove = FALSE
)
}
\arguments{
\item{dat}{data.frame normalize data table}

\item{sif}{data.frame sample information file including model factors}

\item{annot}{data.frame annotation with Symbol column for functional analysis}

\item{species}{character available species: bt ce dr dm gg hs mm pt rn ss xt}

\item{model}{character anova model factors (see details)}

\item{paired}{character factor for paired design}

\item{nested}{character factor for nested design}

\item{batch}{character factor for batch effect design}

\item{addfactor}{character additionnal factors}

\item{qcs}{logical quality controls}

\item{threshold}{numeric vector from 1 to 160 (see details)}

\item{padj}{character fdr by defaut for Benjamini-Hochberg false discovery correction}

\item{pattern}{logical search relevant pattern accross comparisons (see details)}

\item{venn}{logical venn diagram}

\item{cluster}{logical row hierarchical clustering using pearson correlation}

\item{nc}{numeric number of clusters to cut in dendrogramm}

\item{heatmap}{logical do heatmaps for all lists}

\item{maxheatmap}{numeric max rows for heatmap}

\item{volcanoplot}{logical make volcanoplot for each threshold}

\item{lineplot}{logical do lineplot for significant features}

\item{boxplotrow}{logical do boxplot for significant features with Kruskal}

\item{ena}{logical msigdb enrichement analysis (over-representation analysis)}

\item{enamin}{numeric min list size for functional analysis}

\item{filtergeneset}{character regular expression to filter collection geneset (e.g. "reactome|tft")}

\item{enaScoremin}{numeric for minimum ratio ena}

\item{bg}{numeric background used for over-representation test}

\item{sample}{numeric subset analysis}

\item{dopar}{numeric core number}

\item{path}{character results directory path}

\item{dirname}{character results directory name}

\item{zip}{logical compress results directory if TRUE}

\item{remove}{logical remove uncompress results directory if TRUE}
}
\value{
omic results directory
}
\description{
Biostatistics analysis: QCs, ANOVA, threshold filtering, venn diagramm, cluster analysis for significant rows, pattern search

Graphics: volcanoplots, heatmaps, lineplots, boxplots (with Kruskal-Wallis test)

Functional analysis: MSigDB enrichment analysis, stringDB protein interaction network, topGO analysis (gene ontology)

See example section to test workflow with internal data.
}
\details{
Use uem915::env() to load required libraries before uem915::omic() (see example)

Accepted values for threshold param (1 to 150): see list -> uem915:::thresholdlist \%>\% lapply("[",c(1,2)) \%>\% unlist \%>\% matrix(ncol=2,byrow = T) \%>\% data.frame \%>\% setNames(c("pval","fc"))

Pattern: search relevant profiles among up and down comparison combinations

filtergeneset param: see geneset collections -> moalannotgene::genesetdb \%>\% lapply(names)

Experimental design examples:
\itemize{
\item model = "TREAMTENT" for 1-way anova
\item model = "TREATMENT+TIME+TREATMENT*TIME" for 3-ways anova with interaction
\item model = "TREATMENT", paired = "CASE" for 2-ways paired anova
\item model = "TREATMENT", batch = "BATCH" for 2-ways anova with remove batch effect (<=> paired anova)
\item model = "TREATMENT+PHENOTYPE" for 2-ways anova
\item model = "TREATMENT", addfactor = "PHENOTYPE" for 2-ways anova but venn, cluster and pattern are not applied to addfactor
\item model =  "TREATMENT", nested = "TREATMENTinCASE" for 2-ways nested design
\item model =  "TREATMENT", paired = "CASE", batch = "BATCH" for 3-ways paired anova with remove batch effect
}

Limitations:
\itemize{
\item Complete block designs only
\item Use dopar = 2 to decrease computing resources
\item sample param will subsets random rows in dat and decrease analysis time.
}

Annotation updates: 05112023 for gene and ensembl, MSigDB 7.5.1, StringDB 12.0

Input format:
\itemize{
\item Use uem915::input() to load external data from tsv files
\item norm data table must contains IDs in first column
\item norm data (in columns) and sample information (in rows) must have same order
\item norm data and annotation rows must have same order
\item use uem915::annot() function to annotate IDs (Symbols, ensembl and gene ids accepted)
}
}
\examples{
# # test omic() with internal dataset GSE65055:
# # loading libraries
# library(uem915)
# uem915::env()
# # loading norm data
# moal:::GSE65055normdata -> normdata
# normdata \%>\% head
# # loading sample information file
# moal:::GSE65055sampledata -> sampledata
# sampledata \%>\% head
# # ordering factor levels
# sampledata$ANEUPLOIDY \%>\% ordered( c("Control","T13","T18","T21") ) -> sampledata$ANEUPLOIDY
# sampledata$TISSUE \%>\% as.factor -> sampledata$TISSUE
# # annotation
# normdata$rowID \%>\% moal::annot( species = "hs", idtype = "GENE" ) -> annotdata
# # omic analysis
# moal::omic(
#  dat = normdata, sif = sampledata, annot = annotdata, species = "hs",
#  model = "ANEUPLOIDY", batch = "TISSUE", threshold = c(6,126),
#  heatmap = T, lineplot = T, boxplotrow = T,
#  venn = F, cluster = F, pattern = F,
#  ena = T, network = F, topgo = F,
#  sample = NULL, dopar = NULL, zip = F,
#  dirname = "test", path = "." )

}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
