% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.r
\name{qc}
\alias{qc}
\title{Quality Controls}
\usage{
qc(
  dat,
  sif = NULL,
  inputdata = F,
  histo = TRUE,
  boxplot = TRUE,
  hc = TRUE,
  acp = TRUE,
  dirname = NULL,
  path = "."
)
}
\arguments{
\item{dat}{matrix numeric}

\item{sif}{data.frame}

\item{inputdata}{logical to export input data or not}

\item{histo}{logical do histogram if TRUE by defaut}

\item{boxplot}{logical do boxplot if TRUE by defaut}

\item{hc}{logical do hierarchical clustering if TRUE by defaut}

\item{acp}{logical do principal component analysis if TRUE by defaut}

\item{dirname}{character}

\item{path}{character}
}
\value{
data.frame
}
\description{
do descriptive statistics : histogram, boxplot , hierarchical clustering and ACP for column
}
\details{
return pval for each factor of anova model
function use doparalle
}
\examples{
# not run
# qc(dat, metadata)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
