% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacegroupmed.r
\name{replacegroupmed}
\alias{replacegroupmed}
\title{replace value by group median}
\usage{
replacegroupmed(dat, value = 0, factor)
}
\arguments{
\item{dat}{character vector of mzxml file}

\item{value}{numeric value to substitute}

\item{factor}{character}
}
\value{
data.frame
}
\description{
replace value by group median
}
\details{
replace value by column median if group size is one and replace by row group median if group size > 1
}
\examples{
# not run
# replacegroupmed(dat)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
